+++ test/tnep.jl
@@ -9,7 +9,7 @@ end
 
 @testset "test soc tnep" begin
     @testset "3-bus case" begin
-        result = run_tnep("../test/data/case3_tnep.m", SOCWRPowerModel, pajarito_solver; setting = Dict("output" => Dict("branch_flows" => true)))
+        result = run_tnep("../test/data/matpower/case3_tnep.m", SOCWRPowerModel, pajarito_solver; setting = Dict("output" => Dict("branch_flows" => true)))
 
         check_tnep_status(result["solution"])
 
@@ -17,7 +17,7 @@ end
         @test isapprox(result["objective"], 2; atol = 1e-2)
     end
     @testset "5-bus rts case" begin
-        result = run_tnep("../test/data/case5_tnep.m", SOCWRPowerModel, pajarito_solver)
+        result = run_tnep("../test/data/matpower/case5_tnep.m", SOCWRPowerModel, pajarito_solver)
 
         check_tnep_status(result["solution"])
 
@@ -28,7 +28,7 @@ end
 
 @testset "test dc tnep" begin
     @testset "3-bus case" begin
-        result = run_tnep("../test/data/case3_tnep.m", DCPPowerModel, pajarito_solver)
+        result = run_tnep("../test/data/matpower/case3_tnep.m", DCPPowerModel, pajarito_solver)
 
         check_tnep_status(result["solution"])
 
@@ -39,7 +39,7 @@ end
     #=
     # skip this one becouse it is breaking Julia package tests
     @testset "5-bus case" begin
-        result = run_tnep("../test/data/case5_tnep.m", DCPPowerModel, pajarito_solver)
+        result = run_tnep("../test/data/matpower/case5_tnep.m", DCPPowerModel, pajarito_solver)
 
         check_tnep_status(result["solution"])
 
@@ -51,7 +51,7 @@ end
 
 @testset "test dc-losses tnep" begin
     @testset "3-bus case" begin
-        result = run_tnep("../test/data/case3_tnep.m", DCPLLPowerModel, pajarito_solver)
+        result = run_tnep("../test/data/matpower/case3_tnep.m", DCPLLPowerModel, pajarito_solver)
 
         check_tnep_status(result["solution"])
 
@@ -60,7 +60,7 @@ end
     end
 
     @testset "5-bus case" begin
-        result = run_tnep("../test/data/case5_tnep.m", DCPLLPowerModel, pajarito_solver)
+        result = run_tnep("../test/data/matpower/case5_tnep.m", DCPLLPowerModel, pajarito_solver)
 
         check_tnep_status(result["solution"])
 
@@ -74,7 +74,7 @@ if (Pkg.installed("AmplNLWriter") != nothing && Pkg.installed("CoinOptServices")
 
     @testset "test ac tnep" begin
         @testset "5-bus case" begin
-            result = run_tnep("../test/data/case5_tnep.m", ACPPowerModel, BonminNLSolver(["bonmin.bb_log_level=0", "bonmin.nlp_log_level=0"]))
+            result = run_tnep("../test/data/matpower/case5_tnep.m", ACPPowerModel, BonminNLSolver(["bonmin.bb_log_level=0", "bonmin.nlp_log_level=0"]))
 
             check_tnep_status(result["solution"])
 
@@ -88,7 +88,7 @@ end
 
 @testset "test tnep branch flow output" begin
     @testset "3-bus case" begin
-        result = run_tnep("../test/data/case3_tnep.m", SOCWRPowerModel, pajarito_solver; setting = Dict("output" => Dict("branch_flows" => true)))
+        result = run_tnep("../test/data/matpower/case3_tnep.m", SOCWRPowerModel, pajarito_solver; setting = Dict("output" => Dict("branch_flows" => true)))
 
         check_tnep_status(result["solution"])